# Conflicts:
#	packages/cli-plugin-deploy-pulumi/package.json
#	packages/pulumi-aws/package.json
#	packages/pulumi-sdk/src/index.ts
#	yarn.lock

+++ packages/api-file-manager/src/handlers/transform/managers/imageManager.ts
@@ -47,7 +47,14 @@ export default {
             })
             .promise();
 
+        if (!imagesList.Contents) {
+            return;
+        }
+
         for (const imageObject of imagesList.Contents) {
+            if (!imageObject.Key) {
+                continue;
+            }
             await s3.deleteObject(getObjectParams(imageObject.Key)).promise();
         }
     }
